requirements.txt
anicli-ru==4.2.4
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
distlib==0.3.7
filelock==3.12.4
idna==3.4
mutagen==1.47.0
pafy==0.5.5
platformdirs==3.11.0
pycryptodomex==3.19.0
pyfiglet==1.0.2
python-vlc==3.0.18122
requests==2.31.0
urllib3==2.0.7
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
websockets==11.0.3
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz#sha256=edce29ce4d45890af83db6b0fa6d212c5156dffcf1e20e4bc6828e5b57871d84
